Refactor body part classes into seperate files
